project_managing/requirements.txt
amqp==5.3.1
asgiref==3.11.1
attrs==25.4.0
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.5
defusedxml==0.7.1
Django==6.0.2
django-cors-headers==4.9.0
django-filter==25.2
django-redis==6.0.0
django-restframework==0.0.1
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
dotenv==0.9.9
drf-nested-routers==0.95.0
drf-spectacular==0.29.0
flower==2.0.1
gevent==25.9.1
greenlet==3.3.1
gunicorn==25.1.0
humanize==4.15.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
oauthlib==3.3.1
packaging==26.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
redis==7.1.1
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
six==1.17.0
social-auth-app-django==5.7.0
social-auth-core==4.8.5
sqlparse==0.5.5
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
zope.event==6.1
zope.interface==8.2
